v3 proto: reorganized lookup_hash code
